requirements.txt
fastapi
uvicorn
pydantic
motor
pyjwt
requests